Fixes and readme

+++ db/models/categories_models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, String, Date, DateTime
+from sqlalchemy import Column, ForeignKey, String, Date, BigInteger
 from sqlalchemy.orm import relationship
 
 from db.database import Base
@@ -6,10 +6,10 @@ from db.database import Base
 
 class Category(Base):
     __tablename__ = 'categories'
-    id = Column(Integer, primary_key=True, index=True)
+    id = Column(BigInteger, primary_key=True, index=True)
     name = Column(String, index=True)
     creation_date = Column(Date)
 
-    user_id = Column(Integer, ForeignKey('users.id'), index=True)
+    user_id = Column(BigInteger, ForeignKey('users.id'), index=True)
     user = relationship('User',lazy='subquery', back_populates='user_categories')
     expenses = relationship('Expense',lazy='subquery', back_populates='category')
